update delete category in CategoryService

+++ src/main/java/fr/sirine/stock_management_back/service/impl/CategoryService.java
@@ -5,9 +5,11 @@ import fr.sirine.stock_management_back.entities.Category;
 import fr.sirine.stock_management_back.entities.User;
 import fr.sirine.stock_management_back.exceptions.custom.CategoryAlreadyExistException;
 import fr.sirine.stock_management_back.exceptions.custom.CategoryNotFoundException;
+import fr.sirine.stock_management_back.exceptions.custom.IllegalStateException;
 import fr.sirine.stock_management_back.mapper.CategoryMapper;
 import fr.sirine.stock_management_back.repository.CategoryRepository;
 import fr.sirine.stock_management_back.service.ICategoryService;
+import fr.sirine.stock_management_back.service.IProductService;
 import fr.sirine.stock_management_back.service.IUserService;
 import org.springframework.stereotype.Service;
 
@@ -19,11 +21,13 @@ public class CategoryService implements ICategoryService {
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
     private final IUserService userService;
+    private final IProductService productService;
 
-    public CategoryService(CategoryRepository categoryRepository, CategoryMapper categoryMapper, IUserService userService) {
+    public CategoryService(CategoryRepository categoryRepository, CategoryMapper categoryMapper, IUserService userService, IProductService productService) {
         this.categoryRepository = categoryRepository;
         this.categoryMapper = categoryMapper;
         this.userService = userService;
+        this.productService = productService;
     }
 
     public List<CategoryDto> getAllCategories(Integer groupId) {
@@ -40,10 +44,14 @@ public class CategoryService implements ICategoryService {
             categoryRepository.save(category);
         }
     }
-    public void deleteCategory(Integer id) {
+    public void deleteCategory(Integer id, Integer groupId) {
+        categoryRepository.deleteById(id);
+        boolean hasProducts = productService.existsByCategoryIdAndGroupId(id, groupId);
+        if (hasProducts) {
+            throw new IllegalStateException("Impossible de supprimer la catégorie car elle contient des produits.");
+        }
         categoryRepository.deleteById(id);
     }
-
     public Category findById(Integer categoryId) {
         return categoryRepository.findById(categoryId)
                 .orElseThrow(CategoryNotFoundException::new);
